requirements.txt
pandas==2.3.2
numpy==2.3.2
requests==2.32.5
yfinance==1.2.0
python-dotenv==1.2.1
groq==1.1.2
streamlit==1.49.1
plotly==5.22.0
pyarrow==21.0.0
